pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openapi-analyzer"
version = "0.1.0"
description = "A CLI tool for analyzing OpenAPI specifications"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "PyYAML>=6.0",
    "click>=8.0",
    "python-dateutil>=2.8.0",
    "GitPython>=3.1.0",
    "PyGithub>=2.1.0",
]

[project.scripts]
openapi-analyzer = "openapi_analyzer.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
